requirements.txt
tensorflow>=2.10.0
opencv-python>=4.6.0
numpy>=1.21.0
Pillow>=9.0.0
scipy>=1.9.0

